requirements.txt
# ==============================================================================
# Mosaic Framework Dependencies
# Mosaic Framework for LLM Membership Inference Attack
# ==============================================================================

# Core Deep Learning
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Data Processing
datasets>=2.14.0
tokenizers>=0.14.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning & Clustering
scikit-learn>=1.3.0
umap-learn>=0.5.4
hdbscan>=0.8.33

# NLP Tools
nltk>=3.8.0
spacy>=3.6.0
gensim>=4.3.0 # For topic modeling

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Experiment Tracking
tensorboard>=2.14.0
wandb>=0.15.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# API Clients
openai>=1.0.0
anthropic>=0.7.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
rich>=13.5.0
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
